+++ CHANGELOG.md
@@ -3,6 +3,9 @@
 ### Features
 * **noncore**: Ping payload accepts sensor values
 
+### Bug Fixes
+* **core**: only_deleted param of stream endpoint using `deletedAt` instead of `deleted_at`
+
 ## 1.1.6 (2022-05-19)
 
 ### Features