+++ CHANGELOG.md
@@ -4,6 +4,9 @@
 * **noncore**: Ping payload accepts sensor values
 * **noncore**: Segments count endpoint for Guardian diagnostics
 
+### Bug Fixes
+* **core**: only_deleted param of stream endpoint using `deletedAt` instead of `deleted_at`
+
 ## 1.1.6 (2022-05-19)
 
 ### Features