+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-## 1.1.7 (2022-05-??)
+## 1.1.7 (2022-06-06)
 
 ### Features
 * **noncore**: Ping payload accepts sensor values